new command and changed command
Added 'ephemeral', meaning only the user that used the command can see the message (on specific commands).

Added 'activedevbadge' command which replies with a link on how to claim your badge and that you will now be able to in 24hrs.

Split 'user' command to not include token, instead it has its own command ('token') with a warning that no one can do anything with just the start of your token.